+++ RSS-React/.eslintrc.cjs
@@ -12,7 +12,6 @@ module.exports = {
     'plugin:@typescript-eslint/recommended',
     'prettier',
     'plugin:prettier/recommended',
-    'eslint:recommended',
   ],
   overrides: [
     {
